+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @pandacss/cli
 
+## 2.0.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [72580e5]
+  - @pandacss/compiler-shared@2.0.0-beta.8
+  - @pandacss/compiler@2.0.0-beta.8
+  - @pandacss/config@2.0.0-beta.8
+
+## 2.0.0-beta.7
+
+### Patch Changes
+
+- Updated dependencies [97d142a]
+- Updated dependencies [0a11fda]
+  - @pandacss/compiler@2.0.0-beta.7
+  - @pandacss/compiler-shared@2.0.0-beta.7
+  - @pandacss/config@2.0.0-beta.7
+
 ## 2.0.0-beta.6
 
 ### Minor Changes