+++ packages/types/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @pandacss/types
 
+## 2.0.0-beta.8
+
+## 2.0.0-beta.7
+
 ## 2.0.0-beta.6
 
 ### Minor Changes